Prepare logging and debugging
